refactor(types): extract source test modules
- Move type unit tests into sibling *_test.rs files
- Keep shared type definitions separate from test bodies
