+++ crates/aion-types/src/llm.rs
@@ -47,75 +47,5 @@ pub enum LlmEvent {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::message::{StopReason, TokenUsage};
-    use serde_json::json;
-
-    #[test]
-    fn test_thinking_config_enabled_stores_budget() {
-        let config = ThinkingConfig::Enabled { budget_tokens: 4096 };
-        match config {
-            ThinkingConfig::Enabled { budget_tokens } => assert_eq!(budget_tokens, 4096),
-            ThinkingConfig::Disabled => panic!("expected Enabled"),
-        }
-    }
-
-    #[test]
-    fn test_llm_event_text_delta_carries_content() {
-        let event = LlmEvent::TextDelta("hello".to_string());
-        match event {
-            LlmEvent::TextDelta(text) => assert_eq!(text, "hello"),
-            _ => panic!("expected TextDelta"),
-        }
-    }
-
-    #[test]
-    fn test_llm_event_done_carries_stop_reason_and_usage() {
-        let usage = TokenUsage {
-            input_tokens: 10,
-            output_tokens: 20,
-            cache_creation_tokens: 0,
-            cache_read_tokens: 5,
-        };
-        let event = LlmEvent::Done {
-            stop_reason: StopReason::EndTurn,
-            usage,
-        };
-        match event {
-            LlmEvent::Done { stop_reason, usage } => {
-                assert_eq!(stop_reason, StopReason::EndTurn);
-                assert_eq!(usage.input_tokens, 10);
-            }
-            _ => panic!("expected Done"),
-        }
-    }
-
-    #[test]
-    fn test_llm_event_tool_use_fields() {
-        let event = LlmEvent::ToolUse {
-            id: "call_1".to_string(),
-            name: "bash".to_string(),
-            input: json!({"cmd": "ls"}),
-            extra: None,
-        };
-        match &event {
-            LlmEvent::ToolUse { id, name, input, .. } => {
-                assert_eq!(id, "call_1");
-                assert_eq!(name, "bash");
-                assert_eq!(input["cmd"], "ls");
-            }
-            _ => panic!("expected ToolUse"),
-        }
-    }
-
-    #[test]
-    fn test_llm_event_thinking_signature_carries_content() {
-        let event = LlmEvent::ThinkingSignature("sig-123".to_string());
-
-        match event {
-            LlmEvent::ThinkingSignature(signature) => assert_eq!(signature, "sig-123"),
-            _ => panic!("expected ThinkingSignature"),
-        }
-    }
-}
+#[path = "llm_test.rs"]
+mod llm_test;
